added zoom_2d methods to standalone module
